Fix build failure and rename files
